refactor: Decompose AdvancedExpressionFoldingBuilder into multiple classes
